% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coordtrees.R
\name{coordTrees}
\alias{coordTrees}
\title{Get Coordinates of Trees around a center plot}
\usage{
coordTrees(angle, distance, x0 = 0, y0 = 0)
}
\arguments{
\item{angle}{A vector of angle (degrees) of the element with respect to
the center (measured from center).}

\item{distance}{A vector of distance (meters) of the element to the plot
center.}

\item{x0}{x-coordinate of the plot center. Default to 0.}

\item{y0}{y-coordinate of the plot center. Default to 0.}
}
\value{
A \code{dataframe} with two columns (\code{x} and \code{y}) representing
x- and y-coordinates respectively (in meters).
}
\description{
\code{coordTrees} return the x and y coordinates of the trees (o whatever)
located around a center plot.
}
\details{
This function uses distance and angle values of an object to compute the
relative position (coordinates) of the object with respect to coordinates
supplied by \code{x0} and \code{y0}.
}
