% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dendroEco-package.R
\docType{package}
\name{dendroEco-package}
\alias{dendroEco}
\alias{dendroEco-package}
\title{dendroEco: Tools for Computing Dendroecologial Analysis}
\description{
A collection of usefuls functions to compute dendroecological analysis.
}
\author{
\strong{Maintainer}: Antonio J Perez-Luque \email{ajpelu@gmail.com}

}
\keyword{internal}
